Restructure join module in order to remove circular dependency
